+++ book/eslint.config.mjs
@@ -1,4 +1,4 @@
-const { makeConfig } = require('@rockpack/codestyle/index.js');
+import { makeConfig } from '@rockpack/codestyle';
 
 const config = makeConfig();
 
@@ -9,4 +9,4 @@ config.push({
   },
 });
 
-module.exports = config;
+export default config;