+++ e2e/babel-e2e/eslint.config.mjs
@@ -1,4 +1,5 @@
-const { makeConfig } = require('@rockpack/codestyle');
+import { makeConfig } from '@rockpack/codestyle';
+
 const config = makeConfig();
 
 config.push({
@@ -8,4 +9,4 @@ config.push({
   },
 });
 
-module.exports = config;
+export default config;