+++ e2e/babel-e2e/jest.extend.js
@@ -1,0 +1,4 @@
+module.exports = {
+  modulePathIgnorePatterns: ['./src/generators/'],
+  testEnvironment: 'node',
+};