+++ e2e/codestyle-e2e/src/main.spec.ts
@@ -1,5 +1,0 @@
-describe('codestyle generator', () => {
-  it('test', () => {
-    expect(true).toBe(true);
-  });
-});