+++ e2e/starter-e2e/src/generators.spec.ts
@@ -45,7 +45,7 @@ describe('Generators tests', () => {
 
         it('exists lint files', async () => {
           const commitlintrc = await exists(join(genFolder, projectFolder, '.commitlintrc.js'));
-          const eslint = await exists(join(genFolder, projectFolder, 'eslint.config.js'));
+          const eslint = await exists(join(genFolder, projectFolder, 'eslint.config.mjs'));
           const stylelintrc = await exists(join(genFolder, projectFolder, '.stylelintrc.js'));
           expect(commitlintrc).toBeTruthy();
           expect(eslint).toBeTruthy();